Create type.tsx file for Agreement component
