+++ app/components/Agreement/types.tsx
@@ -1,0 +1,13 @@
+import { ChangeEvent } from "react"
+
+export interface AgreementFormProps {
+  terms: {
+    value: boolean[];
+    isValid: boolean;
+  };
+  handleChange: (e: ChangeEvent<HTMLInputElement>) => void;
+}
+
+export interface AgreementInfoProps {
+  disabled: boolean;
+}
